# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant access control and smart building platform API.\n\nAll endpoints are tenant-scoped through the `Authorization: Bearer dm3_live_...`\nor `dm3_test_...` API token. Issue tokens from Settings → API Integration.\nTokens prefixed `dm3_live_` act on production data; `dm3_test_` are sandbox.\n\nThe default rate limit is 60 requests/minute per token. Responses use a\nconsistent envelope: success responses return the resource directly, errors\nreturn `{ \"error\": \"\u003ccode\u003e\", \"message\": \"\u003chuman readable\u003e\" }`.",
        "title": "Duall Master API",
        "termsOfService": "https://duali.com/terms",
        "contact": {
            "name": "Duall Master Support",
            "url": "https://duali.com/support",
            "email": "support@duali.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/attendance/me/leave/requests": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a leave request for the authenticated user. Any caller-supplied user_id is ignored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Submit a leave request",
                "parameters": [
                    {
                        "description": "Leave request (policy_id, start_date, end_date, half_day, reason)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_duali_dm3-backend_pkg_httputil.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_duali_dm3-backend_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_duali_dm3-backend_pkg_httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API token in the form `Bearer dm3_live_...` or `Bearer dm3_test_...`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}